+++ packages/bridge/bridge-shared/src/type.ts
@@ -1,0 +1,9 @@
+export interface ProviderParams {
+  name?: string;
+  basename?: string;
+  memoryRoute?: { entryPath: string };
+}
+
+export interface RenderFnParams extends ProviderParams {
+  dom: HTMLElement;
+}