+++ packages/dts-plugin/src/core/lib/DTSManager.advance.spec.ts
@@ -21,7 +21,7 @@ describe('DTSManager advance usage', () => {
         'react-dom': { singleton: true, eager: true },
       },
     },
-    tsConfigPath: join(__dirname, '../../..', './tsconfig.json'),
+    tsConfigPath: join(__dirname, '../../..', './tsconfig.spec.json'),
     typesFolder: typesFolder,
     compiledTypesFolder: 'compiled-types',
     deleteTypesFolder: false,